runtime: adjust mutex profile samples at recording time
Previously, the contentions and delays in the mutex profile were
adjusted according to the sampling probability at the time when the
profile is reported, based on the most recently configured sampling
rate. However, this makes the profile inaccurate if the mutex profile
fraction is changed at runtime and the profile contains samples recorded
with different rates. As a special case, setting the mutex profile
fraction to 0 would lead to empty mutex profiles being reported.

Instead, adjust the contention count and delay when each sample is
recorded. This eliminates the problem and is consistent with how the
block profiler does it.