runtime: increase CPU profile stack size limit
This CL increases the hard-coded limit for the number of frames in a CPU
profiler call stack to 512 frames. This makes CPU profiles more useful
for programs with deep call stacks, as the previous limit of 64 frames
could often lead to truncation. This limit is still small enough for the
call stack array to fit on the CPU profile signal handler's stack.

Updates #56029

Change-Id: Ib9edfc161b4f8eafe74f81a4df18feed9239e343