runtime/trace: record goroutine labels in execution trace
Goroutines can have key-value labels attached to them through the
runtime/pprof package. These labels are added to samples in the
goroutine and CPU profiles to provide additional context for what the
goroutines were doing when the samples were recorded. However, these
labels are currently not available in execution traces.

This CL adds an new event, traceEvGoroutineLabels, which includes a
goroutine ID and a sequence of key-value pairs from the label set which
was applied to that goroutine. This event is emitted when labels are
updated, and at the start of tracing to record labels that were present
before tracing began.

Fixes #56295.

Change-Id: I2ffdc6fcf76c7f714489361a3b3889833f6c311b