net/http/pprof: add parameter to control delta fields
Add a delta_value query parameter which controls the sample values in a
profile for which deltas are computed when a seconds parameter is given.
This is primarily intended for use with the heap profile. Continuous
profiling services like Google Cloud Profiler or Datadog compute the
"delta" of the alloc_space and alloc_objects profiles over a given
interval, but take the most recent inuse_space and inuse_objects values.
This is done to facilitate computing average values across multiple
profiles. Those services get profiles in-process from the runtime. This
CL gives matching functionality to profilers based on collecting from
the /debug/pprof endpoints, while preserving the previous default
behavior for users who may rely on it.

Change-Id: Ie7a7b9ef9d51f002b62a6ae728362b3b15985fc8